Fix CRCX size_t ref

+++ Utils/CRCX/CRCX.c
@@ -55,7 +55,7 @@ uint8_t CRC8R(uint8_t init, uint8_t poly, const uint8_t * data, uint32_t size)
       return crc;
 }
 
-uint16_t CRC16(uint16_t init, uint16_t poly, const uint8_t * data, size_t size)
+uint16_t CRC16(uint16_t init, uint16_t poly, const uint8_t * data, uint32_t size)
 {
 	uint16_t crc = init;
 	while (size--)
@@ -72,7 +72,7 @@ uint16_t CRC16(uint16_t init, uint16_t poly, const uint8_t * data, size_t size)
 	return crc;
 }
 
-uint16_t CRC16R(uint16_t init, uint16_t poly, const uint8_t * data, size_t size)
+uint16_t CRC16R(uint16_t init, uint16_t poly, const uint8_t * data, uint32_t size)
 {
 	uint16_t crc = init;
 	while (size--)
@@ -90,7 +90,7 @@ uint16_t CRC16R(uint16_t init, uint16_t poly, const uint8_t * data, size_t size)
 }
 
 
-uint16_t CRC16_CCITT(uint16_t init, const uint8_t * data, size_t size)
+uint16_t CRC16_CCITT(uint16_t init, const uint8_t * data, uint32_t size)
 {
 	uint16_t crc = init;
 	while (size--)
